V 1.0

+++ app/src/main/java/com/bernardrenzema/aimm/fragments/OverviewFragment.java
@@ -113,7 +113,7 @@ public class OverviewFragment extends Fragment implements View.OnClickListener, 
                 String barcode = intent != null ? intent.getStringExtra("barcode") : null;
                 String description = intent != null ? intent.getStringExtra("description") : null;
                 String unit = intent != null ? intent.getStringExtra("unit") : null;
-                int quantity = intent != null ? intent.getIntExtra("quantity", 0) : 0;
+                float quantity = intent != null ? intent.getFloatExtra("quantity", 0) : 0;
                 final int type = intent != null ? intent.getIntExtra("type", -1) : -1;
                 final SnackbarHelper snackbarHelper = new SnackbarHelper();
 
@@ -136,7 +136,7 @@ public class OverviewFragment extends Fragment implements View.OnClickListener, 
                                             put("barcode", intent.getStringExtra("barcode")).
                                             put("work", mWork.getWorkId()).
                                             put("client", mClient.getClientId()).
-                                            put("quantity", intent.getIntExtra("quantity", 0));
+                                            put("quantity", intent.getFloatExtra("quantity", 0));
 
                                     switch (type) {
                                         case CreatorActivity.STOCK_ENTRY:
